requirements.txt
Django==4.2.1
django-autoslug==1.9.9
django-extensions==3.2.1
django-formtools==2.4.1
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytils==0.4.1
urllib3==1.26.15
yookassa==2.3.6
